Rename PRE_MIGRATION_LIMIT to ACCOUNT_LIMIT (#5381)
# Motivation

The account limit was put in place to limit the number of accounts
before the migration to stable memory.
But we want to keep the account limit in place to mitigate attacks, even
with accounts in stable memory.

# Changes

Rename `PRE_MIGRATION_LIMIT` to `ACCOUNT_LIMIT` and change the comment
to make sure the limit doesn't get removed as part of the migration
cleanup.

# Tests

Existing tests pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary