Make snsFunctionsStore a derived store (#5437)
# Motivation

Currently, `snsFunctionsStore` is a Svelte store which is filled at the
same time as the `snsAggregatorStore` and based on the same data. This
is fragile and it's more robust to make it a derived store so we have 1
source of truth for the aggregator data.

The same applies to `snsTotalTokenSupplyStore` which we'll do in another
PR.

# Changes

1. Change `snsFunctionsStore` to a derived store.
2. Remove code to fill `snsFunctionsStore`.
3. Remove logic in `sns-ns-functions-project.derived.ts` to fall back on
`snsAggregatorStore` as `snsFunctionsStore` is now derived from
`snsAggregatorStore` so it has the same data.
4. Change import paths where `snsFunctionsStore` is used. 

# Tests

1. Change tests to use `setSnsProjects` instead of setting the
`snsFunctionsStore`, which, being a derived store, can no longer be set
directly.
2. Move `sns-functions.store.spec.ts` to
`sns-functions.derived.spec.ts`.
3. Remove test for the removed fallback logic from
`sns-ns-functions-project.derived.ts`

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary