Fix e2e screenshot name mismatch (#5442)
# Motivation

## Background

1. When an end-to-end test creates a screenshot, it uses the test
name/description in the name of the screenshot file.
2. Our CI test that checks if any screenshots have changed, only looks
at existing screenshots and does not detect newly created screenshots.
3. This means that when https://github.com/dfinity/nns-dapp/pull/4271
changed the description of some tests, instead of comparing to existing
screenshots, it created new screenshots, which did not cause the test to
fail. And since the new screenshots were never committed, the test
effectively stopped testing the screenshots.

## In this PR

We restore the test names that were changed in
https://github.com/dfinity/nns-dapp/pull/4271, as well as the behavior
of those tests, to make the screenshot as close as possible to what it
would have been before the test stopped working.

I did not bring back the replacement of tokens names and symbols
because:
1. An attribute that was used to do that is no longer there.
2. It's no longer necessary because the SNS names are now deterministic.

## Not in this PR

We should also fix the CI test to detect new screenshot but that will be
done in another PR.

# Changes

1. Change `"Accounts"` back to `"My Tokens"` in test names.
3. Navigate to the tokens table.
4. Update screenshots.

# Tests

test only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary