Detect new screenshots (#5438)
# Motivation

We had the bug in the screenshot e2e test. If a test created new
screenshot, those were not detected and did not cause a test failure.

# Changes

1. Fix the CI script to also look at new screenshot files.
2. Add a missing Firefox screenshot which was previously only added for
Chrome, which was detected by the fixed script in [this
run](https://github.com/dfinity/nns-dapp/actions/runs/10826399700/job/30037362884).

# Tests

test only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary