Fix a bug where the perf counter is blown away in init (#5456)
# Motivation

The perf counter was recorded right before getting wiped by replacing
the state, so `init set_canister_arguments` is never visible through
`get_stats`. Although, this is not very important for mainnet since we
don't expect init to execute there.

# Changes

* Similar to `post_upgrade`, the counter is saved before the state
initialization, but recorded (to the state) after initialization
* Additional changes to make `init` more consistent with `post_upgrade`

# Tests

N/A

# Todos

- [x] Add entry to changelog (if necessary).