Simplify the lifecycle and access points of `State` (#5449)
# Motivation

The initialization of the `STATE` currently is rather complicated - it
had a `State::default()` before reading from the stable memory, and
there is some logic to replace it with the correct value, which makes it
difficult to reason about the value of each of its field. Instead,
making `STATE` itself a `RefCell<Option<State>>` seems simpler - there
is a `Some(State)` or None.

Each field in `State` does not need `Default`, which is helpful because
not all objects have reasonable `Default` (like `Partitions` and
`AccountStore`)

# Changes

* Encapsulate `STATE` within `state.rs` and define its accessors
* Replace `STATE.with(...)` with those accessors
* Remove the `Default` of `State`
* Change default->replace to initializing `State` directly in the
post_upgrade

# Tests

N/A

# Todos

- [x] Add entry to changelog (if necessary).