Fix flakiness in e2e/sns-governance.spec.ts (#5472)
# Motivation

I noticed some flakiness in `e2e/sns-governance.spec.ts` because it was
waiting for skeleton cards to disappear while the neurons table we have
now has a spinner instead of skeleton cards.
This was already fixed for NNS in
https://github.com/dfinity/nns-dapp/commit/2395a228b10930469e1f3165e67bf7fc5faeff64#diff-34fb0aee6c9d4c7899b1a1d5a88c149c53c205559da5a91eb977e2798203500a
but we somehow skipped SNS.

# Changes

1. Wait for the spinner to go away in `waitForContentLoaded`.
2. Remove irrelevant code in the PO related to skeleton cards.
3. Drive-by: Make `stakeFirstSnsNeuron` consistent with
`stakeFirstNnsNeuron` by waiting for the staking view to disappear after
staking the SNS neuron.

# Tests

Without the change, `e2e/sns-governance.spec.ts` failed 2 out of 12
times. With the change I ran it 24 times without failure.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary