Move SNS modals tests (#5473)
# Motivation

For a component defined in file
`frontend/src/lib/components/XXX/Foo.svelte` the test file is usually in
`frontend/src/tests/lib/components/XXX/Foo.svelte` for the same `XXX`.
So when this is not the case, it can be confusing when trying to run the
tests for a component.

But for the components in
`frontend/src/lib/modals/sns/{neurons,sale,proposals}/` the tests were
all lumped together in `frontend/src/tests/lib/modals/sns/`.

# Changes

1. Create directories `neurons`, `sale`, and `proposals` under
`frontend/src/tests/lib/modals/sns/`.
2. Move the tests in the appropriate directories.

# Tests

Tests only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary