Check for minimum stake when staking SNS tokens (#5491)
# Motivation

We recently had an issue where the `SnsStakeNeuronModal` failed to check
the minimum stake amount, which resulted in some people trying to stake
too little.

In this PR I add an extra check in the service to make sure we don't try
to stake too little.
Providing a proper error message requires passing in more information
about the token and i18n so I didn't do this.
It's still the responsibility of the component calling the service to
check this, but now we have an extra safety check.

# Changes

1. When staking SNS tokens, check that the staked amount is at least the
minimum stake.

# Tests

1. Update some tests that were accidentally affected by this.
2. Add a test with a staked amount that's too small.
3. Tested manually after temporarily removing the check in the
component.
4. Tested manually that increasing an existing stake by a small amount
is still possible.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary