Fix checksum for ICRC-1 addresses (#5496)
# Motivation

The textual address format of an ICRC-1 account with subaccount has
leading zeros removed from the subaccount.
The address also contains a checksum of the account and subaccount but
when calculating the checksum, the leading zeros should not be removed
from the subaccount.
The `convert-id` script did this wrong.

# Changes

Calculate the checksum with the subaccount without leading zeros
removed.

# Tests

Test updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary