Update remove imported dialog props to support principal only (#5504)
# Motivation

In the case of a failed imported token, we don’t have the token
metadata, such as the logo and token name, but users should still be
able to remove the imported token. To achieve this, we first need to
make the confirmation dialog work with the ledger canister ID only.

# Changes

- Change the confirmation component props to work seamlessly with either
the universe or the ledger canister ID.

# Tests

Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.