Redirection on unknown token. (#5528)
# Motivation

When the user is on the imported token wallet page and the sign-out
timer expires, the page reloads, leaving the user on an empty page. This
happens because imported tokens are available only when the user is
signed in. Upon sign-out, we no longer know the type of the token
(whether it’s ICRC or not). Rather than making guesses, it’s more
reliable to navigate the user to the main tokens page in this case.

# Changes

- Navigate to the tokens table when it’s not possible to identify the
token. Since this logic is not intended to be reused, it is implemented
directly in the component rather than in a derived store.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.